+++ packages/quantic/force-app/main/default/lwc/quanticSearchBox/quanticSearchBox.js
@@ -1,6 +1,11 @@
-import { LightningElement, api, track } from 'lwc';
+import {LightningElement, api, track} from 'lwc';
 // @ts-ignore
-import { registerComponentForInit, initializeWithHeadless } from 'c/quanticHeadlessLoader';
+import {
+  registerComponentForInit,
+  initializeWithHeadless,
+} from 'c/quanticHeadlessLoader';
+
+import search from '@salesforce/label/c.quantic_Search';
 
 const ENTER = 13;
 const ARROWUP = 38;
@@ -14,43 +19,72 @@ export default class QuanticSearchBox extends LightningElement {
     suggestions: [],
     value: '',
   };
-
   /** @type {any} */
   get suggestions() {
-    return this.state.suggestions.map((s, index) => ({ key: index, value: s.rawValue }));
+    return this.state.suggestions.map((s, index) => ({
+      key: index,
+      value: s.highlightedValue,
+    }));
   }
 
+  labels = {
+    search
+  }
+  
   /** @type {string} */
   @api engineId;
   /** @type {string} */
-  @api placeholder = 'Search';
+  @api placeholder = `${this.labels.search}...`;
+  /** @type {boolean} */
+  @api withoutSubmitButton = false;
+  /** @type {number} */
+  @api numberOfSuggestions = 5;
 
+
+  /** @type {string} */
+  searchBoxContainerClass;
   /** @type {import("coveo").SearchBox} */
   searchBox;
   /** @type {import("coveo").Unsubscribe} */
   unsubscribe;
+
   /** @type {number} */
   selectionIndex = -1;
   /** @type {HTMLInputElement} */
   input;
-  /** @type {import("lwc").HTMLElementTheGoodPart} */
+  /** @type {HTMLElement} */
   combobox;
-
+  /** @type {HTMLButtonElement} */
+  clearButton;
   /** @type {() => void} */
-  resetSelectionIndex = () => {this.selectionIndex = -1};
+  resetSelectionIndex = () => {
+    this.selectionIndex = -1;
+  };
   /** @type {() => boolean} */
-  // @ts-ignore
-  areSuggestionsShown = () => (this.template.querySelector('.slds-combobox').classList.contains('slds-is-open'));
-  /** @type {() => HTMLElement[]} */
-  // @ts-ignore
-  getSuggestionElements = () => this.template.querySelectorAll('.slds-listbox__option');
+  areSuggestionsShown = () =>
+    this.template
+      .querySelector('.slds-combobox')
+      .classList.contains('slds-is-open');
+  /** @type {() => NodeListOf<HTMLElement>} */
+  getSuggestionElements = () =>
+    this.template.querySelectorAll('.slds-listbox__option');
 
   /**
    * @param {import("coveo").SearchEngine} engine
    */
   @api
   initialize(engine) {
-    this.searchBox = CoveoHeadless.buildSearchBox(engine);
+    this.searchBox = CoveoHeadless.buildSearchBox(engine, {
+      options: {
+        numberOfSuggestions: this.numberOfSuggestions,
+        highlightOptions: {
+          notMatchDelimiters: {
+            open: '<b>',
+            close: '</b>',
+          },
+        },
+      },
+    });
     this.unsubscribe = this.searchBox.subscribe(() => this.updateState());
   }
 
@@ -61,13 +95,15 @@ export default class QuanticSearchBox extends LightningElement {
   renderedCallback() {
     initializeWithHeadless(this, this.engineId, this.initialize.bind(this));
     if (!this.input) {
-      // @ts-ignore
       this.input = this.template.querySelector('input');
     }
     if (!this.combobox) {
-      // @ts-ignore
       this.combobox = this.template.querySelector('.slds-combobox');
     }
+    if (!this.clearButton) {
+      this.clearButton = this.template.querySelector('.slds-button__icon');
+    }
+    this.setSearchBoxContainerClass();
   }
 
   disconnectedCallback() {
@@ -83,14 +119,24 @@ export default class QuanticSearchBox extends LightningElement {
     this.state = this.searchBox.state;
   }
 
+  setSearchBoxContainerClass() {
+    if(this.withoutSubmitButton){
+      this.searchBoxContainerClass = "slds-combobox__form-element slds-input-has-icon slds-input-has-icon_left-right";
+      this.input.setAttribute("aria-labelledby", "fixed-text-label")
+    } else{
+      this.searchBoxContainerClass = "slds-combobox__form-element slds-input-has-icon slds-input-has-icon_right slds-input-has-fixed-addon";
+      this.input.setAttribute("aria-labelledby", "fixed-text-label fixed-text-addon-post");
+    }
+  }
+
   showSuggestions() {
     this.combobox.classList.add('slds-is-open');
-    this.combobox.setAttribute("aria-expanded", true);
+    this.combobox.setAttribute('aria-expanded', 'true');
   }
 
   hideSuggestions() {
     this.combobox.classList.remove('slds-is-open');
-    this.combobox.setAttribute('aria-expanded', false);
+    this.combobox.setAttribute('aria-expanded', 'false');
     this.resetHighlighted();
     this.resetSelectionIndex();
   }
@@ -105,7 +151,7 @@ export default class QuanticSearchBox extends LightningElement {
     this.resetHighlighted();
     suggestion.setAttribute('aria-selected', 'true');
     suggestion.classList.add('slds-has-focus');
-    this.input.value = suggestion.textContent;
+    this.input.value = suggestion.innerText;
   }
 
   resetHighlighted() {
@@ -149,6 +195,12 @@ export default class QuanticSearchBox extends LightningElement {
     this.setHighlighted();
   }
 
+  onSubmit(){
+    this.searchBox.updateText(this.input.value);
+    this.searchBox.submit();
+    this.input.blur();
+  }
+
   /**
    * @param {KeyboardEvent & {target: {value : string}}} event
    */
@@ -167,24 +219,30 @@ export default class QuanticSearchBox extends LightningElement {
   }
 
   onFocus() {
+    this.clearButton.classList.remove('slds-hidden');
+    this.clearButton.classList.add('slds-visible');
     this.searchBox.showSuggestions();
     this.showSuggestions();
   }
 
   onBlur() {
+    this.clearButton.classList.remove('slds-visible');
+    this.clearButton.classList.add('slds-hidden');
     this.hideSuggestions();
+  }
+
+  clearInput() {
+    this.input.value = '';
+    this.searchBox.updateText(this.input.value);
+    this.input.focus();
   }
 
   preventDefault(event) {
     event.preventDefault();
   }
 
-  /**
-   * @param {KeyboardEvent & {target: {textContent : string}}} event
-   */
   handleSuggestionSelection(event) {
-    const textValue = event.target.textContent;
-
+    const textValue = event.target.innerText;
     this.updateSearchboxText(textValue);
     this.searchBox.submit();
     this.input.blur();